# feat: use avx2