--- a/arith_generic.go
+++ b/arith_generic.go
@@ -1,0 +1,35 @@
+//go:build !amd64 || purego
+
+package uint256
+
+func (z *Int) And(x, y *Int) *Int {
+	z[0] = x[0] & y[0]
+	z[1] = x[1] & y[1]
+	z[2] = x[2] & y[2]
+	z[3] = x[3] & y[3]
+	return z
+}
+
+func (z *Int) Or(x, y *Int) *Int {
+	z[0] = x[0] | y[0]
+	z[1] = x[1] | y[1]
+	z[2] = x[2] | y[2]
+	z[3] = x[3] | y[3]
+	return z
+}
+
+func (z *Int) Xor(x, y *Int) *Int {
+	z[0] = x[0] ^ y[0]
+	z[1] = x[1] ^ y[1]
+	z[2] = x[2] ^ y[2]
+	z[3] = x[3] ^ y[3]
+	return z
+}
+
+func (z *Int) Not(x *Int) *Int {
+	z[0] = ^x[0]
+	z[1] = ^x[1]
+	z[2] = ^x[2]
+	z[3] = ^x[3]
+	return z
+}