--- a/cpu_amd64.go
+++ b/cpu_amd64.go
@@ -1,0 +1,11 @@
+//go:build amd64 && !purego
+
+package uint256
+
+import (
+	"golang.org/x/sys/cpu"
+)
+
+var (
+	hasAVX2 = cpu.X86.HasAVX2
+)