--- a/cpu_generic.go
+++ b/cpu_generic.go
@@ -1,0 +1,7 @@
+//go:build !amd64 || purego
+
+package uint256
+
+const (
+	hasAVX2 = false
+)