--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,5 @@
 module github.com/holiman/uint256
 
 go 1.19
+
+require golang.org/x/sys v0.25.0