--- a/uint256.go
+++ b/uint256.go
@@ -953,12 +953,6 @@ func (z *Int) srsh192(x *Int) {
 	z[3], z[2], z[1], z[0] = math.MaxUint64, math.MaxUint64, math.MaxUint64, x[3]
 }
 
-// Not sets z = ^x and returns z.
-func (z *Int) Not(x *Int) *Int {
-	z[3], z[2], z[1], z[0] = ^x[3], ^x[2], ^x[1], ^x[0]
-	return z
-}
-
 // Gt returns true if z > x
 func (z *Int) Gt(x *Int) bool {
 	return x.Lt(z)
@@ -1238,33 +1232,6 @@ func (z *Int) ISRsh(n uint) *Int {
 // Set sets z to x and returns z.
 func (z *Int) Set(x *Int) *Int {
 	z[0], z[1], z[2], z[3] = x[0], x[1], x[2], x[3]
-	return z
-}
-
-// Or sets z = x | y and returns z.
-func (z *Int) Or(x, y *Int) *Int {
-	z[0] = x[0] | y[0]
-	z[1] = x[1] | y[1]
-	z[2] = x[2] | y[2]
-	z[3] = x[3] | y[3]
-	return z
-}
-
-// And sets z = x & y and returns z.
-func (z *Int) And(x, y *Int) *Int {
-	z[0] = x[0] & y[0]
-	z[1] = x[1] & y[1]
-	z[2] = x[2] & y[2]
-	z[3] = x[3] & y[3]
-	return z
-}
-
-// Xor sets z = x ^ y and returns z.
-func (z *Int) Xor(x, y *Int) *Int {
-	z[0] = x[0] ^ y[0]
-	z[1] = x[1] ^ y[1]
-	z[2] = x[2] ^ y[2]
-	z[3] = x[3] ^ y[3]
 	return z
 }