use avx2 in amd64 may has some perf improvement.