@cuiweixie 2025-09-24
> Strangely enough, I am getting a regression in speeds on this. And yes, my system does have avx2 (I checked that it enters the desired path).
> 
> ```
> [user@work uint256]$ benchstat ops.before ops.after 
> goos: linux
> goarch: amd64
> pkg: github.com/holiman/uint256
> cpu: 12th Gen Intel(R) Core(TM) i7-1270P
>                      │  ops.before  │              ops.after               │
>                      │    sec/op    │    sec/op     vs base                │
> And/single/uint256-8   3.858n ± 19%   6.852n ± 33%  +77.60% (p=0.000 n=10)
> Or/single/uint256-8    3.735n ± 21%   4.727n ± 49%  +26.58% (p=0.005 n=10)
> Xor/single/uint256-8   4.512n ± 32%   6.167n ± 34%  +36.67% (p=0.014 n=10)
> geomean                4.021n         5.845n        +45.38%
> ```
> 
> The one odd thing about my system is that I'm using qubes os, so I am executing inside a virtual machine. I don't really think that should matter, but who knows.
> 
> What do you get, if you benchmark e.g. like this, on master versus this pr?
> 
> ```
> go test . -run - -bench "And/single/u|Or/single/u|Xor/single/u" --count 10
> ```

```
goos: darwin
goarch: amd64
pkg: github.com/holiman/uint256
cpu: VirtualApple @ 2.50GHz
                      │   old.txt    │                new.txt                │
                      │    sec/op    │    sec/op     vs base                 │
And/single/uint256-10   1.127n ±  2%   2.455n ± 32%  +117.93% (p=0.000 n=10)
Or/single/uint256-10    1.104n ±  2%   2.400n ±  1%  +117.49% (p=0.000 n=10)
Xor/single/uint256-10   1.245n ± 40%   2.395n ±  3%   +92.37% (p=0.000 n=10)
geomean                 1.157n         2.417n        +108.91%
```
I think we can close this pr now. I just test only one benchmark by goland to check the it can improve  the perf.
It may be caused by some noise. 1ns level is easy to make some mistake.

Apologies for taking up your time.